nixos/containers: Reload container service when possible
By removing the container `path` from all service attributes, which cause a
restart during a `switch-to-configuration`, the container service can be
reloaded (using `switch-to-configuration test`) when only the system `path`
changes.

The separation is achieved by introducing a new configuration file, which
contains all the settings that don't affect the host or the container setup
process.